fix profile picture component
